rewriting tobybot in kotlin.
misc commands working

+++ src/main/kotlin/toby/jpa/persistence/IExcusePersistence.kt
@@ -7,8 +7,8 @@ interface IExcusePersistence {
     fun listApprovedGuildExcuses(guildId: Long?): List<ExcuseDto?>
     fun listPendingGuildExcuses(guildId: Long?): List<ExcuseDto?>
     fun createNewExcuse(excuseDto: ExcuseDto?): ExcuseDto?
-    fun getExcuseById(id: Int?): ExcuseDto
+    fun getExcuseById(id: Long?): ExcuseDto
     fun updateExcuse(excuseDto: ExcuseDto): ExcuseDto
     fun deleteAllExcusesForGuild(guildId: Long?)
-    fun deleteExcuseById(id: Int?)
+    fun deleteExcuseById(id: Long?)
 }
